Move containers from containers/thing/index.jsx to containers/thing/container.jsx - Karl's feedback
